+++ src/graphql/schemas/Vote.js
@@ -36,6 +36,11 @@ export default `
     voted: Boolean
     voteResults: VoteResult
   }
+
+  type VoteStatistic {
+    proceduresCount: Int
+    votedProcedures: Int
+  }
   
   type Mutation {
     vote(procedure: ID!, selection: VoteSelection!): Vote
@@ -43,5 +48,6 @@ export default `
   
   type Query {
     votes(procedure: ID!): Vote
+    voteStatistic: VoteStatistic
   }
   `;