+++ src/services/sms/index.js
@@ -1,4 +1,6 @@
 import request from 'request';
+import _ from 'lodash';
+
 import CONSTANTS from '../../config/constants';
 
 export const statusSMS = async SMSID => {
@@ -9,6 +11,10 @@ export const statusSMS = async SMSID => {
   };
 
   return new Promise((resolve, reject) => {
+    if (CONSTANTS.SMS_SIMULATE) {
+      resolve(true);
+      return;
+    }
     request({ url, qs }, (err, response, body) => {
       if (err) {
         Log.error(JSON.stringify(err));
@@ -138,6 +144,10 @@ export const sendSMS = async (phone, code) => {
   };
 
   return new Promise((resolve, reject) => {
+    if (CONSTANTS.SMS_SIMULATE) {
+      resolve({ status: true, statusCode: 100, SMSID: _.uniqueId() });
+      return;
+    }
     request({ url, qs }, (err, response, body) => {
       let status = false;
       let statusCode = null;